Handle multiple microchip OIDs.
We've realized microchip has 3 OIDs. This handles the 2 OIDs we've seen in
production - we don't want to allocate the third unless we start shipping chips
with that OID.

The change also changes the GET_UID command to return the Ja Rule Model & the
MAC address. We'll use this in the production tool.